add generic base strategy issue #105
