+++ src/Grace/DependencyInjection/Impl/EnumerableStrategies/ImmutableLinkListStrategy.cs
@@ -18,20 +18,7 @@ namespace Grace.DependencyInjection.Impl.EnumerableStrategies
         public ImmutableLinkListStrategy(IInjectionScope injectionScope) : base(typeof(ImmutableLinkedList<>), injectionScope)
         {
         }
-
-        /// <summary>
-        /// Get an activation expression for this strategy
-        /// </summary>
-        /// <param name="scope"></param>
-        /// <param name="request"></param>
-        /// <param name="lifestyle"></param>
-        /// <returns></returns>
-        public override IActivationExpressionResult GetDecoratorActivationExpression(IInjectionScope scope, IActivationExpressionRequest request,
-            ICompiledLifestyle lifestyle)
-        {
-            throw new NotSupportedException("Decorators on collection is not supported at this time");
-        }
-
+        
         /// <summary>
         /// Get an activation expression for this strategy
         /// </summary>
